Refactoring. Naming CardList component correctly.
